requirements.txt
numpy>=1.14.0
pandas>=0.23.4
gdown>=3.10.1
tqdm>=4.30.0
Pillow>=5.2.0
opencv-python==4.4.0.46
tensorflow>=1.9.0
keras>=2.2.0